cordillera: group frame vocabulary under an `Http2` object
The frame-layer types had deliberately generic names — `Frame`, `Flags`,
`Setting`, `ErrorCode`, `FrameType`, `SettingId` — that shouldn't occupy the
top-level `soundness.*` namespace. Consolidate the five files into one
`Http2.scala` nesting them under `object Http2` (mirroring how telekinesis
groups its types under `Http`), and export just `Http2`. Internal consumers use
`import Http2.*`. No behavioural change; 30 tests, 0 warnings.
