+++ lib/cordillera/src/core/soundness_cordillera_core.scala
@@ -32,6 +32,5 @@
                                                                                                   */
 package soundness
 
-export cordillera.{Frame, FrameType, Flags, ErrorCode, Setting, SettingId, Hpack, HpackEntry,
-    HpackTable, Huffman, H2Connection, H2Stream, Http2Error, FrameReader, PseudoHeaders, H2Endpoint,
-    Http2Client}
+export cordillera.{Http2, Hpack, HpackEntry, HpackTable, Huffman, H2Connection, H2Stream,
+    Http2Error, FrameReader, PseudoHeaders, H2Endpoint, Http2Client}
